Commit 02/05/2022

+++ src/main/java/com/rentalkars/servlet/RentController.java
@@ -1,6 +1,7 @@
 package com.rentalkars.servlet;
 
 import com.rentalkars.hibernate.dao.RentDao;
+import com.rentalkars.hibernate.entity.Car;
 import com.rentalkars.hibernate.entity.Rent;
 
 import javax.servlet.*;
@@ -15,18 +16,14 @@ import java.util.List;
 public class RentController extends HttpServlet {
 
     private final RentDao rDao = new RentDao();
-    private Rent rent = null;
     private RequestDispatcher rd;
-
-    private Long rentId;
-    private LocalDate startDate;
-    private LocalDate endDate;
 
     private String errorMsg;
 
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+
         String command = request.getParameter("command");
         if(command == null) {
             command = "LIST";
@@ -57,86 +54,125 @@ public class RentController extends HttpServlet {
                 listRents(request, response);
                 break;
         }
+
     }
 
     private void listRents(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+
         List<Rent> rentList = rDao.getRents();
 
         request.setAttribute("rentsList", rentList);
 
         rd = request.getRequestDispatcher("/sections/reservations_list.jsp");
         rd.forward(request, response);
+
     }
 
 
     private void checkAvailable(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        startDate = LocalDate.parse(request.getParameter("startDate"));
-        endDate = LocalDate.parse(request.getParameter("endDate"));
 
-        checkBefore(endDate, startDate, request, response);
-        checkBefore(startDate, LocalDate.now(), request, response);
+        LocalDate startDate = LocalDate.parse(request.getParameter("startDate"));
+        LocalDate endDate = LocalDate.parse(request.getParameter("endDate"));
 
-        rent = new Rent(startDate, endDate);
+        if(endDate.isBefore(startDate) || startDate.isBefore(LocalDate.now())) {
+
+            errorMsg = "La data finale inserita non è valida. Riprova";
+            request.setAttribute("errorMsg", errorMsg);
+            rd = request.getRequestDispatcher("/user/manage_reservations.jsp");
+            rd.forward(request, response);
+
+        } else {
+
+
+            Car car = new Car();
+            Rent rent = new Rent(startDate, endDate, car);
+            addReservation(request, response, rent);
+
+
+        }
+
+        //listRents(request, response);
+
+    }
+
+    public void addReservation(HttpServletRequest request, HttpServletResponse response, Rent rent) throws ServletException, IOException {
+
 
         rDao.saveRent(rent);
         listRents(request, response);
 
     }
 
-    private void checkBefore (LocalDate d1, LocalDate d2, HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        if(d1.isBefore(d2)) {
-            errorMsg = "La data finale inserita non è valida. Riprova";
-            request.setAttribute("errorMsg", errorMsg);
-            rd = request.getRequestDispatcher("/user/manage_reservations.jsp");
-            rd.forward(request, response);
-        }
-    }
-
     private void updateReservation(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        rentId = Long.valueOf(request.getParameter("rentId"));
-        startDate = LocalDate.parse(request.getParameter("startDate"));
-        endDate = LocalDate.parse(request.getParameter("endDate"));
 
-        rent = rDao.selById(rentId);
+        Long rentId = Long.valueOf(request.getParameter("rentId"));
+        LocalDate startDate = LocalDate.parse(request.getParameter("startDate"));
+        LocalDate endDate = LocalDate.parse(request.getParameter("endDate"));
 
-        Period period = Period.between(startDate, rent.getStartDate());
+        Rent rent = rDao.selById(rentId);
 
-        if(period.getYears() == 0 && period.getMonths() == 0 && period.getDays() <= 2) {
+        if(checkTime(LocalDate.now().until(startDate))) {
+
             errorMsg = "Non è possibile modificare la prenotazione. La data è troppo vicina";
+            request.setAttribute("rentUpdate", rent);
             request.setAttribute("errorMsg", errorMsg);
             rd = request.getRequestDispatcher("/user/manage_reservations.jsp");
             rd.forward(request, response);
-        }
+
+        } else {
+
+            rent.setStartDate(startDate);
+            rent.setEndDate(endDate);
             rDao.updateReservation(rent);
             listRents(request, response);
 
+        }
+
+    }
+
+
+    public boolean checkTime(Period timeDiff) {
+        return (timeDiff.getYears() == 0 && timeDiff.getMonths() == 0 && timeDiff.getDays() < 2);
     }
 
 
     private void loadReservation(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        rentId = Long.valueOf(request.getParameter("rentId"));
 
-        rent = rDao.selById(rentId);
+        Long rentId = Long.valueOf(request.getParameter("rentId"));
 
-        request.setAttribute("rentUpdate", rent);
-        rd = request.getRequestDispatcher("/admin/manage_reservations.jsp");
-        rd.forward(request, response);
+        Rent rent = rDao.selById(rentId);
+
+        if(checkTime(LocalDate.now().until(rent.getStartDate()))) {
+
+            errorMsg = "Non è possibile modificare la prenotazione. La data è troppo vicina";
+            request.setAttribute("errorMsg", errorMsg);
+
+        } else {
+            request.setAttribute("rentUpdate", rent);
+            rd = request.getRequestDispatcher("/user/manage_reservations.jsp");
+            rd.forward(request, response);
+        }
+
+        listRents(request, response);
+
     }
 
 
     private void deleteReservation(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        rentId = Long.valueOf(request.getParameter("rentId"));
 
-        rent = rDao.selById(rentId);
+        Long rentId = Long.valueOf(request.getParameter("rentId"));
+        Rent rent = rDao.selById(rentId);
 
-        rDao.removeReservation(rent);
+        if(checkTime(LocalDate.now().until(rent.getStartDate()))) {
 
+            errorMsg = "Non è possibile cancellare la prenotazione. La data è troppo vicina";
+            request.setAttribute("errorMsg", errorMsg);
+
+        } else {
+            rDao.removeReservation(rent);
+        }
         listRents(request, response);
-    }
-
-
-    @Override
-    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 
     }
+
 }
